test(meta):fix pkg name from meta to meta_test.
