+++ meta/config_test.go
@@ -1,13 +1,14 @@
-package meta
+package meta_test
 
 import (
 	"github.com/BurntSushi/toml"
+	"github.com/cnosdb/cnosdb/meta"
 	"testing"
 )
 
 func TestConfig_Parse(t *testing.T) {
 	// Parse configuration.
-	var c Config
+	var c meta.Config
 	if _, err := toml.Decode(`
 dir = "/tmp/foo"
 hostname = "localhost"
