+++ meta/data_test.go
@@ -1,134 +1,31 @@
-package meta
+package meta_test
 
 import (
-	"fmt"
 	"github.com/cnosdb/cnosdb"
+	"github.com/cnosdb/cnosdb/meta"
 	"github.com/cnosdb/cnosdb/vend/cnosql"
-	"github.com/cnosdb/cnosdb/vend/db/pkg/testing/assert"
 	"reflect"
-	"sort"
 	"testing"
 	"time"
 )
 
-func TestShardGroupSort_Shard_EndTimeNotEqual(t *testing.T) {
-	sg1 := ShardGroupInfo{
-		ID:          1,
-		StartTime:   time.Unix(1000, 0),
-		EndTime:     time.Unix(1100, 0),
-		TruncatedAt: time.Unix(1050, 0),
-	}
-
-	sg2 := ShardGroupInfo{
-		ID:        2,
-		StartTime: time.Unix(1000, 0),
-		EndTime:   time.Unix(1100, 0),
-	}
-
-	sgs := ShardGroupInfos{sg2, sg1}
-
-	sort.Sort(sgs)
-
-	if sgs[len(sgs)-1].ID != 2 {
-		t.Fatal("unstable sort for ShardGroupInfos")
-	}
-}
-
-func TestShardGroupSort_Shard_EndTimeEqual(t *testing.T) {
-	sg1 := ShardGroupInfo{
-		ID:        1,
-		StartTime: time.Unix(900, 0),
-		EndTime:   time.Unix(1100, 0),
-	}
-
-	sg2 := ShardGroupInfo{
-		ID:        2,
-		StartTime: time.Unix(1000, 0),
-		EndTime:   time.Unix(1100, 0),
-	}
-
-	sgs := ShardGroupInfos{sg2, sg1}
-
-	sort.Sort(sgs)
-
-	if sgs[len(sgs)-1].ID != 2 {
-		t.Fatal("unstable sort for ShardGroupInfos")
-	}
-}
-
-func TestShardGroupInfo_Contains(t *testing.T) {
-	sgi := &ShardGroupInfo{StartTime: time.Unix(10, 0), EndTime: time.Unix(20, 0)}
-
-	tests := []struct {
-		ts  time.Time
-		exp bool
-	}{
-		{time.Unix(0, 0), false},
-		{time.Unix(9, 0), false},
-		{time.Unix(10, 0), true},
-		{time.Unix(11, 0), true},
-		{time.Unix(15, 0), true},
-		{time.Unix(19, 0), true},
-		{time.Unix(20, 0), false},
-		{time.Unix(21, 0), false},
-	}
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("ts=%d", test.ts.Unix()), func(t *testing.T) {
-			got := sgi.Contains(test.ts)
-			assert.Equal(t, got, test.exp)
-		})
-	}
-}
-
-func Test_Data_RetentionPolicy_MarshalBinary(t *testing.T) {
-	zeroTime := time.Time{}
-	epoch := time.Unix(0, 0).UTC()
-
-	startTime := zeroTime
-	sgi := &ShardGroupInfo{
-		StartTime: startTime,
-	}
-	isgi := sgi.marshal()
-	sgi.unmarshal(isgi)
-	if got, exp := sgi.StartTime.UTC(), epoch.UTC(); got != exp {
-		t.Errorf("unexpected start time.  got: %s, exp: %s", got, exp)
-	}
-
-	startTime = time.Unix(0, 0)
-	endTime := startTime.Add(time.Hour * 24)
-	sgi = &ShardGroupInfo{
-		StartTime: startTime,
-		EndTime:   endTime,
-	}
-	isgi = sgi.marshal()
-	sgi.unmarshal(isgi)
-	if got, exp := sgi.StartTime.UTC(), startTime.UTC(); got != exp {
-		t.Errorf("unexpected start time.  got: %s, exp: %s", got, exp)
-	}
-	if got, exp := sgi.EndTime.UTC(), endTime.UTC(); got != exp {
-		t.Errorf("unexpected end time.  got: %s, exp: %s", got, exp)
-	}
-	if got, exp := sgi.DeletedAt.UTC(), zeroTime.UTC(); got != exp {
-		t.Errorf("unexpected DeletedAt time.  got: %s, exp: %s", got, exp)
-	}
-}
-
 func Test_Data_DropDatabase(t *testing.T) {
-	data := &Data{
-		Databases: []DatabaseInfo{
+
+	data := &meta.Data{
+		Databases: []meta.DatabaseInfo{
 			{Name: "db0"},
 			{Name: "db1"},
 			{Name: "db2"},
 			{Name: "db4"},
 			{Name: "db5"},
 		},
-		Users: []UserInfo{
+		Users: []meta.UserInfo{
 			{Name: "user1", Privileges: map[string]cnosql.Privilege{"db1": cnosql.ReadPrivilege, "db2": cnosql.WritePrivilege}},
 			{Name: "user2", Privileges: map[string]cnosql.Privilege{"db2": cnosql.WritePrivilege}},
 		},
 	}
 
-	expDbs := make([]DatabaseInfo, 4)
+	expDbs := make([]meta.DatabaseInfo, 4)
 	copy(expDbs, data.Databases[1:])
 	if err := data.DropDatabase("db0"); err != nil {
 		t.Fatal(err)
@@ -136,21 +33,21 @@ func Test_Data_DropDatabase(t *testing.T) {
 		t.Fatalf("got %v, expected %v", got, exp)
 	}
 
-	expDbs = []DatabaseInfo{{Name: "db1"}, {Name: "db2"}, {Name: "db5"}}
+	expDbs = []meta.DatabaseInfo{{Name: "db1"}, {Name: "db2"}, {Name: "db5"}}
 	if err := data.DropDatabase("db4"); err != nil {
 		t.Fatal(err)
 	} else if got, exp := data.Databases, expDbs; !reflect.DeepEqual(got, exp) {
 		t.Fatalf("got %v, expected %v", got, exp)
 	}
 
-	expDbs = []DatabaseInfo{{Name: "db1"}, {Name: "db2"}}
+	expDbs = []meta.DatabaseInfo{{Name: "db1"}, {Name: "db2"}}
 	if err := data.DropDatabase("db5"); err != nil {
 		t.Fatal(err)
 	} else if got, exp := data.Databases, expDbs; !reflect.DeepEqual(got, exp) {
 		t.Fatalf("got %v, expected %v", got, exp)
 	}
 
-	expUsers := []UserInfo{
+	expUsers := []meta.UserInfo{
 		{Name: "user1", Privileges: map[string]cnosql.Privilege{"db1": cnosql.ReadPrivilege}},
 		{Name: "user2", Privileges: map[string]cnosql.Privilege{}},
 	}
@@ -162,14 +59,14 @@ func Test_Data_DropDatabase(t *testing.T) {
 }
 
 func Test_Data_CreateRetentionPolicy(t *testing.T) {
-	data := Data{}
+	data := meta.Data{}
 
 	err := data.CreateDatabase("foo")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	err = data.CreateRetentionPolicy("foo", &RetentionPolicyInfo{
+	err = data.CreateRetentionPolicy("foo", &meta.RetentionPolicyInfo{
 		Name:     "bar",
 		ReplicaN: 1,
 		Duration: 24 * time.Hour,
@@ -188,27 +85,27 @@ func Test_Data_CreateRetentionPolicy(t *testing.T) {
 	}
 
 	// Try to recreate the same RP with default set to true, should fail
-	err = data.CreateRetentionPolicy("foo", &RetentionPolicyInfo{
+	err = data.CreateRetentionPolicy("foo", &meta.RetentionPolicyInfo{
 		Name:     "bar",
 		ReplicaN: 1,
 		Duration: 24 * time.Hour,
 	}, true)
-	if err == nil || err != ErrRetentionPolicyConflict {
-		t.Fatalf("unexpected error.  got: %v, exp: %s", err, ErrRetentionPolicyConflict)
+	if err == nil || err != meta.ErrRetentionPolicyConflict {
+		t.Fatalf("unexpected error.  got: %v, exp: %s", err, meta.ErrRetentionPolicyConflict)
 	}
 
 	//Try to recreate the RP with the same name, should fail
-	err = data.CreateRetentionPolicy("foo", &RetentionPolicyInfo{
+	err = data.CreateRetentionPolicy("foo", &meta.RetentionPolicyInfo{
 		Name:     "bar",
 		ReplicaN: 1,
 		Duration: 12 * time.Hour,
 	}, false)
-	if err == nil || err != ErrRetentionPolicyExists {
-		t.Fatalf("unexpected error.  got: %v, exp: %s", err, ErrRetentionPolicyExists)
+	if err == nil || err != meta.ErrRetentionPolicyExists {
+		t.Fatalf("unexpected error.  got: %v, exp: %s", err, meta.ErrRetentionPolicyExists)
 	}
 
 	// Creating the same RP with the same specifications should succeed
-	err = data.CreateRetentionPolicy("foo", &RetentionPolicyInfo{
+	err = data.CreateRetentionPolicy("foo", &meta.RetentionPolicyInfo{
 		Name:     "bar",
 		ReplicaN: 1,
 		Duration: 24 * time.Hour,
@@ -219,7 +116,7 @@ func Test_Data_CreateRetentionPolicy(t *testing.T) {
 }
 
 func TestData_AdminUserExists(t *testing.T) {
-	data := Data{}
+	data := meta.Data{}
 
 	// No users means no admin.
 	if data.AdminUserExists() {
@@ -297,7 +194,7 @@ func TestData_AdminUserExists(t *testing.T) {
 }
 
 func TestData_SetPrivilege(t *testing.T) {
-	data := Data{}
+	data := meta.Data{}
 	if err := data.CreateDatabase("db0"); err != nil {
 		t.Fatal(err)
 	}
@@ -307,7 +204,7 @@ func TestData_SetPrivilege(t *testing.T) {
 	}
 
 	// When the user does not exist, SetPrivilege returns an error.
-	if got, exp := data.SetPrivilege("not a user", "db0", cnosql.AllPrivileges), ErrUserNotFound; got != exp {
+	if got, exp := data.SetPrivilege("not a user", "db0", cnosql.AllPrivileges), meta.ErrUserNotFound; got != exp {
 		t.Fatalf("got %v, expected %v", got, exp)
 	}
 
@@ -323,7 +220,7 @@ func TestData_SetPrivilege(t *testing.T) {
 }
 
 func TestData_TruncateShardGroups(t *testing.T) {
-	data := &Data{}
+	data := &meta.Data{}
 
 	must := func(err error) {
 		if err != nil {
@@ -332,7 +229,7 @@ func TestData_TruncateShardGroups(t *testing.T) {
 	}
 
 	must(data.CreateDatabase("db"))
-	rp := NewRetentionPolicyInfo("rp")
+	rp := meta.NewRetentionPolicyInfo("rp")
 	rp.ShardGroupDuration = 24 * time.Hour
 	must(data.CreateRetentionPolicy("db", rp, true))
 
@@ -421,7 +318,7 @@ func TestData_TruncateShardGroups(t *testing.T) {
 }
 
 func TestUserInfo_AuthorizeDatabase(t *testing.T) {
-	emptyUser := &UserInfo{}
+	emptyUser := &meta.UserInfo{}
 	if !emptyUser.AuthorizeDatabase(cnosql.NoPrivileges, "anydb") {
 		t.Fatal("expected NoPrivileges to be authorized but it wasn't")
 	}
@@ -435,22 +332,8 @@ func TestUserInfo_AuthorizeDatabase(t *testing.T) {
 		t.Fatal("expected AllPrivileges to prevent authorization, but it was authorized")
 	}
 
-	adminUser := &UserInfo{Admin: true}
+	adminUser := &meta.UserInfo{Admin: true}
 	if !adminUser.AuthorizeDatabase(cnosql.AllPrivileges, "anydb") {
 		t.Fatalf("expected admin to be authorized but it wasn't")
-	}
-}
-
-func TestNodeInfo_serializes(t *testing.T) {
-	node1 := &NodeInfo{1, "localhost", "127.0.0.1"}
-
-	info := node1.marshal()
-	if info == nil {
-		t.Fatalf("marshal failed")
-	}
-	node2 := &NodeInfo{}
-	node2.unmarshal(info)
-	if !reflect.DeepEqual(node1, node2) {
-		t.Fatalf("unmarshal failed")
 	}
 }
